+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Change Log
 
-Version 0.19.0 *(In development)*
----------------------------------
+Version 0.19.0 *(2024-05-17)*
+-----------------------------
+
+- Bug fix: Fix EmojiUtils.isOnlyEmojis not detecting emojis with variant selectors. [\#1087](https://github.com/vanniktech/Emoji/pull/1087) ([vanniktech](https://github.com/vanniktech))
+- API: New emoji-androidx-emoji2 module which uses androidx.emoji2:emoji2 [\#1086](https://github.com/vanniktech/Emoji/pull/1086) ([vanniktech](https://github.com/vanniktech))
+- Technical: Update Emoji source to 15.1.2 to include new emojis. [\#1085](https://github.com/vanniktech/Emoji/pull/1085) ([vanniktech](https://github.com/vanniktech))
 
 Version 0.18.0 *(2023-12-11)*
 -----------------------------